% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p.R
\name{calc_p}
\alias{calc_p}
\title{Efficient calcuation of regularized empirical inverse quantiles for each data point.}
\usage{
calc_p(x, MX = 1)
}
\arguments{
\item{x}{sorted values}

\item{MX}{-- regulariztion limit}
}
\description{
Efficient calcuation of regularized empirical inverse quantiles for each data point.
}
