% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt.R
\name{estimate}
\alias{estimate}
\title{Estimate WGT transformation for data.}
\usage{
estimate(x, theta = NULL, eta = NULL)
}
\arguments{
\item{x}{-- data vector to be transformed}

\item{theta}{-- percentage along geodesic according to Wasserstein distance}

\item{eta}{-- percentage along geodesic according to normalized linearly-invariate Wasserstein distance}
}
\description{
Estimate WGT transformation for data.
}
